Customize windows ad

+++ providers/domain.rb
@@ -110,13 +110,14 @@ action :join do
             Add-Computer -DomainName #{new_resource.name} -Credential $mycreds -Force:$true -Restart
           EOH
         else
-          code "netdom join #{node[:hostname]} /d #{new_resource.name} /ud:#{new_resource.domain_user} /pd:#{new_resource.domain_pass} /reboot"
+          code "netdom join #{node[:hostname]} /d #{new_resource.name} /ud:#{new_resource.domain_user} /pd:#{new_resource.domain_pass} /ou: \'#{new_resource.ou}\' /reboot"  
+		  
         end
       end
 
     new_resource.updated_by_last_action(false)
     end
-
+	
     new_resource.updated_by_last_action(true)
   end
 end
@@ -124,13 +125,17 @@ end
 action :unjoin do
   if computer_exists?
     powershell_script "unjoin_#{new_resource.name}" do
-      code <<-EOH
-      $secpasswd = ConvertTo-SecureString '#{new_resource.domain_pass}' -AsPlainText -Force
-      $mycreds = New-Object System.Management.Automation.PSCredential ('#{new_resource.domain_user}', $secpasswd)
-      Remove-Computer -UnjoinDomainCredential $mycreds -Force:$true -Restart
-      EOH
+	  if node[:os_version] >= "6.2"
+		  code <<-EOH
+		  $secpasswd = ConvertTo-SecureString '#{new_resource.domain_pass}' -AsPlainText -Force
+		  $mycreds = New-Object System.Management.Automation.PSCredential ('#{new_resource.domain_user}', $secpasswd)
+		  Remove-Computer -UnjoinDomainCredential $mycreds -Force:$true -Restart
+		  EOH
+      else
+		  code "netdom remove /d:#{new_resource.name} #{node[:hostname]} /ud:#{new_resource.domain_user} /pd:#{new_resource.domain_pass} /reboot"
+	  end
     end
-
+	
     new_resource.updated_by_last_action(true)
   else
     Chef::Log.debug("The computer is already a member of a workgroup")
